# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Anvesh API",
    "version": "1.0.0",
    "description": "Anvesh is a local-first travel discovery platform for India.\n\nEvery response uses one envelope. Success is `{ success: true, data, meta }`;\nfailure is `{ success: false, error: { code, message, details? }, meta }`.\n`meta.requestId` matches the `X-Request-Id` header and the server log line.\n\nMoney is always an integer number of minor units (paise). Coordinates are\nGeoJSON `[longitude, latitude]`.\n\nTokens are portal-scoped: a token minted for one portal presented to another\nis rejected with `PORTAL_MISMATCH`.",
    "license": {
      "name": "Proprietary"
    }
  },
  "servers": [
    {
      "url": "http://localhost:4000/api/v1",
      "description": "Local development"
    },
    {
      "url": "https://api.anvesh.travel/api/v1",
      "description": "Production"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "SuccessEnvelope": {
        "type": "object",
        "required": [
          "success",
          "data",
          "meta"
        ],
        "properties": {
          "success": {
            "type": "boolean",
            "enum": [
              true
            ]
          },
          "data": {
            "type": "object"
          },
          "meta": {
            "type": "object",
            "required": [
              "requestId"
            ],
            "properties": {
              "requestId": {
                "type": "string"
              }
            }
          }
        }
      },
      "ErrorEnvelope": {
        "type": "object",
        "required": [
          "success",
          "error",
          "meta"
        ],
        "properties": {
          "success": {
            "type": "boolean",
            "enum": [
              false
            ]
          },
          "error": {
            "type": "object",
            "required": [
              "code",
              "message"
            ],
            "properties": {
              "code": {
                "type": "string",
                "enum": [
                  "INTERNAL_ERROR",
                  "NOT_FOUND",
                  "BAD_REQUEST",
                  "VALIDATION_ERROR",
                  "INVALID_INPUT",
                  "CONFLICT",
                  "RATE_LIMITED",
                  "SERVICE_UNAVAILABLE",
                  "PAYLOAD_TOO_LARGE",
                  "UNSUPPORTED_MEDIA_TYPE",
                  "IDEMPOTENCY_KEY_REQUIRED",
                  "IDEMPOTENCY_KEY_REUSED",
                  "UNAUTHORIZED",
                  "AUTH_INVALID_CREDENTIALS",
                  "AUTH_TOKEN_EXPIRED",
                  "AUTH_TOKEN_INVALID",
                  "AUTH_TOKEN_REUSED",
                  "AUTH_TOKEN_REVOKED",
                  "AUTH_EMAIL_NOT_VERIFIED",
                  "AUTH_EMAIL_ALREADY_REGISTERED",
                  "AUTH_ACCOUNT_SUSPENDED",
                  "AUTH_ACCOUNT_PENDING",
                  "AUTH_RESET_TOKEN_INVALID",
                  "AUTH_VERIFICATION_TOKEN_INVALID",
                  "AUTH_PASSWORD_TOO_WEAK",
                  "AUTH_TOTP_REQUIRED",
                  "AUTH_TOTP_INVALID",
                  "AUTH_TOTP_ALREADY_ENABLED",
                  "AUTH_TOTP_NOT_ENABLED",
                  "AUTH_LOGIN_LOCKED",
                  "FORBIDDEN",
                  "PORTAL_MISMATCH",
                  "PORTAL_NOT_ALLOWED",
                  "ROLE_NOT_ALLOWED",
                  "NOT_RESOURCE_OWNER",
                  "ADMIN_INVITE_INVALID",
                  "ADMIN_INVITE_EXPIRED",
                  "ADMIN_INVITE_ALREADY_USED",
                  "USER_NOT_FOUND",
                  "GUIDE_NOT_FOUND",
                  "GUIDE_PROFILE_INCOMPLETE",
                  "GUIDE_NOT_VERIFIED",
                  "PLACE_NOT_FOUND",
                  "PLACE_NOT_PUBLISHED",
                  "PLACE_SLUG_TAKEN",
                  "EXPERIENCE_NOT_FOUND",
                  "EXPERIENCE_NOT_PUBLISHED",
                  "DESTINATION_NOT_FOUND",
                  "CATEGORY_NOT_FOUND",
                  "COLLECTION_NOT_FOUND",
                  "COLLECTION_LIMIT_REACHED",
                  "MODERATION_INVALID_TRANSITION",
                  "REVIEW_NOT_FOUND",
                  "REVIEW_ALREADY_EXISTS",
                  "REVIEW_NOT_ELIGIBLE",
                  "REVIEW_EDIT_WINDOW_CLOSED",
                  "SLOT_NOT_FOUND",
                  "SLOT_CLOSED",
                  "SLOT_IN_PAST",
                  "SLOT_SOLD_OUT",
                  "SLOT_OVERLAP",
                  "SLOT_HAS_BOOKINGS",
                  "BOOKING_NOT_FOUND",
                  "BOOKING_INVALID_STATE",
                  "BOOKING_EXPIRED",
                  "BOOKING_NOT_CANCELLABLE",
                  "BOOKING_SEATS_INVALID",
                  "PAYMENT_NOT_FOUND",
                  "PAYMENT_PROVIDER_NOT_CONFIGURED",
                  "PAYMENT_PROVIDER_ERROR",
                  "PAYMENT_SIGNATURE_INVALID",
                  "PAYMENT_ALREADY_CAPTURED",
                  "PAYMENT_AMOUNT_MISMATCH",
                  "PAYMENT_INVALID_STATE",
                  "WEBHOOK_SIGNATURE_INVALID",
                  "REFUND_NOT_ALLOWED",
                  "REFUND_AMOUNT_INVALID",
                  "TRIP_NOT_FOUND",
                  "TRIP_DAY_NOT_FOUND",
                  "TRIP_ACTIVITY_NOT_FOUND",
                  "TRIP_LIMIT_REACHED",
                  "AI_PROVIDER_NOT_CONFIGURED",
                  "AI_PROVIDER_ERROR",
                  "AI_OUTPUT_INVALID",
                  "AI_HALLUCINATED_REFERENCE",
                  "AI_QUOTA_EXCEEDED",
                  "AI_PROMPT_REJECTED",
                  "MEDIA_NOT_FOUND",
                  "MEDIA_TYPE_NOT_ALLOWED",
                  "MEDIA_TOO_LARGE",
                  "MEDIA_UPLOAD_INCOMPLETE",
                  "MEDIA_STORAGE_NOT_CONFIGURED",
                  "MEDIA_QUOTA_EXCEEDED",
                  "MAPS_PROVIDER_NOT_CONFIGURED",
                  "MAPS_PROVIDER_ERROR",
                  "NOTIFICATION_NOT_FOUND",
                  "REPORT_NOT_FOUND",
                  "REPORT_ALREADY_RESOLVED",
                  "RECOMMENDATION_CONFIG_NOT_FOUND",
                  "RECOMMENDATION_CONFIG_INVALID"
                ]
              },
              "message": {
                "type": "string"
              },
              "details": {
                "type": "object"
              }
            }
          },
          "meta": {
            "type": "object",
            "required": [
              "requestId"
            ],
            "properties": {
              "requestId": {
                "type": "string"
              }
            }
          }
        }
      }
    },
    "parameters": {
      "RequestId": {
        "name": "X-Request-Id",
        "in": "header",
        "required": false,
        "schema": {
          "type": "string"
        },
        "description": "Correlation id. Generated when absent and echoed on every response."
      }
    }
  },
  "paths": {
...
    "/api/v1/notifications": {
      "get": {
        "summary": "Your notification feed",
        "tags": [
          "Notifications"
        ],
        "operationId": "get-notifications",
        "parameters": [
          {
            "$ref": "#/components/parameters/RequestId"
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid or expired token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "Role not allowed for this operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "Validation failed (VALIDATION_ERROR)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "429": {
            "description": "Rate limited (RATE_LIMITED)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected failure (INTERNAL_ERROR)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/notifications/{id}/read": {
      "post": {
        "summary": "Mark one notification read",
        "tags": [
          "Notifications"
        ],
        "operationId": "post-notifications-by-id-read",
        "parameters": [
          {
            "$ref": "#/components/parameters/RequestId"
          },
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "description": "Validated with the shared Zod schema in @anvesh/validation. Unknown fields are rejected."
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid or expired token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "Role not allowed for this operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "Validation failed (VALIDATION_ERROR)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "429": {
            "description": "Rate limited (RATE_LIMITED)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected failure (INTERNAL_ERROR)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/notifications/read-all": {
      "post": {
        "summary": "Mark every notification read",
        "tags": [
          "Notifications"
        ],
        "operationId": "post-notifications-read-all",
        "parameters": [
          {
            "$ref": "#/components/parameters/RequestId"
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "description": "Validated with the shared Zod schema in @anvesh/validation. Unknown fields are rejected."
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid or expired token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "Role not allowed for this operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "Validation failed (VALIDATION_ERROR)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "429": {
            "description": "Rate limited (RATE_LIMITED)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected failure (INTERNAL_ERROR)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
...
  }
}